Avoid double/float conversions

+++ Eris/Entity.cpp
@@ -205,11 +205,11 @@ void Entity::updatePredictedState(const WFMath::TimeStamp& t)
 {
     assert(isMoving());
     
-    double dt = (t - m_lastMoveTime).milliseconds() / 1000.0; 
+    float dt = (t - m_lastMoveTime).milliseconds() / 1000.0f; 
     
     if (m_acc.isValid()) {
         m_predicted.velocity = m_velocity + (m_acc * dt);
-        m_predicted.position = m_position + (m_velocity * dt) + (m_acc * 0.5 * dt * dt);
+        m_predicted.position = m_position + (m_velocity * dt) + (m_acc * 0.5f * dt * dt);
     } else {
         m_predicted.velocity = m_velocity;
         m_predicted.position = m_position + (m_velocity * dt);
